Make GetStringHeight return an unsigned int
A height is a distance and distances are unsigned; stop pretending
otherwise in GetStringHeight. This fixes initialisation of two
variables in GetStringMultiLineBoundingBox with clang and helps
avoid some casts in GameOptionsWindow::UpdateWidgetSize.

+++ src/settings_gui.cpp
@@ -382,7 +382,7 @@ struct GameOptionsWindow : Window {
 				/* Find the biggest description for the default size. */
 				for (int i = 0; i < BaseGraphics::GetNumSets(); i++) {
 					SetDParamStr(0, BaseGraphics::GetSet(i)->GetDescription(GetCurrentLanguageIsoCode()));
-					size->height = max(size->height, (uint)GetStringHeight(STR_BLACK_RAW_STRING, size->width));
+					size->height = max(size->height, GetStringHeight(STR_BLACK_RAW_STRING, size->width));
 				}
 				break;
 
@@ -401,7 +401,7 @@ struct GameOptionsWindow : Window {
 				/* Find the biggest description for the default size. */
 				for (int i = 0; i < BaseSounds::GetNumSets(); i++) {
 					SetDParamStr(0, BaseSounds::GetSet(i)->GetDescription(GetCurrentLanguageIsoCode()));
-					size->height = max(size->height, (uint)GetStringHeight(STR_BLACK_RAW_STRING, size->width));
+					size->height = max(size->height, GetStringHeight(STR_BLACK_RAW_STRING, size->width));
 				}
 				break;
 
@@ -409,7 +409,7 @@ struct GameOptionsWindow : Window {
 				/* Find the biggest description for the default size. */
 				for (int i = 0; i < BaseMusic::GetNumSets(); i++) {
 					SetDParamStr(0, BaseMusic::GetSet(i)->GetDescription(GetCurrentLanguageIsoCode()));
-					size->height = max(size->height, (uint)GetStringHeight(STR_BLACK_RAW_STRING, size->width));
+					size->height = max(size->height, GetStringHeight(STR_BLACK_RAW_STRING, size->width));
 				}
 				break;
